chore: Normalize test names
